server/requirements.txt
aniso8601==9.0.1
anytree==2.12.1
asgiref==3.8.1
Django==4.2.15
django-cors-headers==4.4.0
django-filter==21.1
django-graphql-jwt==0.4.0
graphene==3.3
graphene-django==3.2.2
graphene-django-filter==0.6.5
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
stringcase==1.2.0
text-unidecode==1.3
typing_extensions==4.10.0
tzdata==2024.1
wrapt==1.16.0
